Commint 20-10-2017

+++ CerberusMultiBranch/Support/Constants.cs
@@ -11,7 +11,7 @@ namespace CerberusMultiBranch.Support
         public const int One = 1;
         public const int Two = 2;
 
-        public const int OneHundred = 100;
+        public const double OneHundred = 100;
 
         public const int QuickResults = 300;
